+++ src/main/java/dta/sfmflow/datagen/ModLanguageProvider.java
@@ -54,6 +54,7 @@ public class ModLanguageProvider extends LanguageProvider {
 		add("gui.sfmflow.item_conditional", "Item Conditional");		
 		add("gui.sfmflow.fluid_conditional", "Fluid Conditional");
 		add("gui.sfmflow.energy_conditional", "Energy Conditional");
+		add("gui.sfmflow.redstone_conditional", "Redstone Conditional");
 		
 		add("gui.sfmflow.observer_trigger", "Observer Trigger");